Refresh auth token at the beginning of 'cf push'
This is bad and we should feel bad. This is not the solution you're looking for.
The real solution would be to have the backend (in this case, CC) be able to
verify that a request's access token is good while the application bits are
still being sent. Barring that, we could also refresh the auth token and try
uploading again, but that seems like OVERKILL.

[Finishes #69034628]


Former-commit-id: 7e75def62f3fd5985607e62599584b159c1e9826