Work around SAML configuration in UAA
... provides a `--sso` flag to `cf login`.

This is purposefully hardcoded today so that the behavior for login with
username and password will continue to be simple, and the behavior for SSO
with a UAA server configured for SAML will necessarily need to be driven out.

What I like least about this personally is that we're requiring our users
to understand HOW they should authenticate to the authentication server,
when ideally the authentication server should be telling us how users should
authenticate.

My apologies for the code duplication.

[Finishes #69963402, #69963432]


Former-commit-id: f239cea753003361d3a6cd03894f95a841c8253d