Start removing more global state
It's been more than a little disappointing that we've had assertions in
our tests that rely on a global bool in a testhelper package. Changing
this for one test now so that we have a pattern to abide by later. My
hope is that we can root out the rest of these one by one.

[Finishes #68922816]


Former-commit-id: 21d26f97d03923d276f95dc4c44c34bb06c44474