Simplify ui FailWithUsage method
It was absolutely absurd that we decided we needed to duplicate the
name of each command so that we could pass it to the UI object to
print usage. Maybe this change is only possible now because
codegangsta/cli updated its interface in the last few months, but
I'm still disappoint.

Yet another amazing perl find and replace.

[Finishes #70898678]


Former-commit-id: fdd6a80e782e8027ae2050b08e69bfba3cb7a1e7