Add more instructions for first time checkout
We removed the instructions for godep. Upon further reflection, we
didn't want to duplicate their instructions in case they change in the
future.

Added a note that you can safely ignore a warning when 'go get'ing the
cli. This warning is caused by the cli not having any .go files at the
root.

Added instructions to cd to the directory for the cli before adding a
remote. I, Tim Jarratt, had assumed that people would know how to do
this, but I think it's helpful to provide a little more context to
people coming to this for the first time.

[Finishes #70686778]


Former-commit-id: d04de17984c8548c1963667dd6dea15f26508323