Add support for reading environment for an app
The biggest change here is that we are also expecting to return the
VCAP_SERVICES hash. Realistically we'll only ever get this if the change
that the services team recently made to CloudController is actually
present.

We believe that returning these two objects separately is the best bet
for now. The VCAP_SERVICES hash will be pretty printed, but the user
hash will not be (for the time being).

[Finishes #71250896]


Former-commit-id: 7b4b5c291540280d5fc0711761fe3d3b1caa3349