Clean up temp directories in i18n
Whoops. Caught this on CI. Thought about testing it, but it does not
seem like the sort of behavior that is easy to test since the way this
fails is that the OS fails to create new tmp dirs. Faking out the temp
directory creation is ... let us just say down that path lies madness.


Former-commit-id: 815f3d3febb2c49670ee5e4d51db5d3fd19f7162