Pivotal Tap
Previously, the CLI could be installed via the central homebrew
repository.  Recently the restrictions placed on the formula in that
became too onerous to continue hosting it there.  The formula is now
hosted in the Pivotal Tap[1].  This change updates the documentation
with instructions on how to add the Pivotal Tap and install the CLI
from it.

[1]: https://github.com/pivotal/homebrew-tap


Former-commit-id: 8102aac6287e466975083c62dfc7c192fe62c8b5