Decreate startup timeout in start_test.go
We feel rather strongly that increasing these timeouts is not really a
solution for making tests pass or fail. A better pattern would be to
apply a clock function as in the gateway or log_message_queue tests.

That would make these tests easier to maintain, faster and also less
flakey.


Former-commit-id: b2e59c266c7a9b053cc9856542b65ac30b00074f