Table printer accepts var args in Add(strings...)
Pretty useful because a lot of the time you just have some strings
you want to print, and not a slice. Realistically this should probably
be typed to only accept a number of strings equal to the headers
provided but that's an improvement for another day. The big win here
is reducing the amount of cruft we have to write when we use this
interface.

[#72663452]


Former-commit-id: 2c016114866546c7f1b79c79fea56147c9607b97