Make bin/test more robust against ginkgo upgrades breaking the build
We found that when we upgrade ginkgo in our godeps, our tests tend to go
red and stay red for a while (typically until someone will ssh into a
box and poke around until it's fixed). The solution tends to involve

* upgrading the ginkgo binary
* removing the .a files in the $GOPATH/pkg directory

The previous commit assures we run the ginkgo binary from godep. This
commit ensures we clean out the .a files from the previous test run.


Former-commit-id: c6e3a70dbf49d110cfda75d4a6c5155d56bde048