Generate FakeUserRepository with counterfeiter
This is mostly a direct conversion. Only minimal work was done to the
tests to take advantage of what the new fake can do; many old tests
were missing assertions about arguments to methods or to the exact
number of methods. Some of these were updated, many were not. It is very
possible that many tests are still not where they need to be.


Former-commit-id: 3b01b789e6b64728ecc0920572b9e326affa8261