Generate FakeUserRepository with counterfeiter
This is mostly a direct conversion. Only minimal work was done to the
tests to take advantage of what the new fake can do; many old tests
were missing assertions about arguments to methods or to the exact
number of methods. Some of these were updated, many were not. It is very
possible that many tests are still not where they need to be.


Former-commit-id: 07c182674cd6c79f1b56811261832a750ad519c1 [formerly 07c182674cd6c79f1b56811261832a750ad519c1 [formerly 3b01b789e6b64728ecc0920572b9e326affa8261]]
Former-commit-id: 2b9ac22cec3d40b9b25d6a7b2d49714a22c0cef3
Former-commit-id: 473740c7a0f0d6d0f12280a73680c9d421bbc1b6