Move api-switching logic into Requirements phase
There are two flags that need to be checked for; set_roles_by_username
and unset_roles_by_username.

This necessitates moving the logic out of the UserRequirement and back
into the command, but higher up than it was before.

Presently UnsetSpaceRole passes 'true' always in the call to
NewUserRequirement. This is in accordance with how it worked previously
and will be updated in a future story.

[#97841328]

Signed-off-by: Simon Leung <leungs@us.ibm.com>
Signed-off-by: Kris Hicks <khicks@pivotal.io>

Former-commit-id: 30e30286320a189f80b4a006b09139ab4dc983e2 [formerly 30e30286320a189f80b4a006b09139ab4dc983e2 [formerly 92ae3eeacb074bc6a83b2b275aa035479bcd0e48]]
Former-commit-id: 5c3255c5164ee9f5098d7811d4501a5dc3b873a4
Former-commit-id: 76d33b74da830fb7da0f664369fcdd4fd1baf010