Add optimization to WalkAppFiles
filepath.Walk gives you the option to stop traversing a directory by
returning filepath.SkipDir from your WalkFunc.

When filepath.Walk gives us a directory that should be ignored, we will:

* If there was no error, return filepath.SkipDir to stop traversal of
  the dir
* If there was an error and the OS is Windows, try harder (see below)
* If there was an error and the OS is not Windows, just return the
  original error

On Windows filepath.Walk will give your WalkFunc:

* the path it tried to traverse
* a nil os.FileInfo
* an error

when it tries to traverse a directory with a path that is beyond its
MAX_LENGTH. We can work around MAX_LENGTH by prepending `\\?\` to the
path and doing an os.Lstat with that path.

The code for Windows platforms will be tested via GATS on Windows.

[#102327200]


Former-commit-id: b973fa044681ec0adb088f63a0a4c0cf1a4c6371 [formerly b973fa044681ec0adb088f63a0a4c0cf1a4c6371 [formerly 6b80f70fec349846e132f746ee74447a0ce766d5]]
Former-commit-id: af13167ecaad7539dffcabe7364f83dfc45d964c
Former-commit-id: f2874939dfb49c648bdfdfd18de2b607c41e1fc9