Remove 'app' local var
This var could be misleading; it's only used to get the GUID from the
app we retrieved by name via the application requirement.

From that point on we need the more fully-fledged app we get from
AppSummaryRepository.

Having two locals with similar names has led to issues in the past,
where one is used for some things and one is used for others, but for
seeminly no reason, leading to hard-to-find bugs.


Former-commit-id: b4ec7915df18fbdad8f234c8ef69cdf97a2b07cc [formerly b4ec7915df18fbdad8f234c8ef69cdf97a2b07cc [formerly e9b36bbe574b572324f5be3e88522196baff6267]]
Former-commit-id: 649c470802dae3039fcfd660417c0a510a0724aa
Former-commit-id: e7daf9de421489a412e4ffc5de4ed5286c241494