roll back remove panic from app command
It turns out that command tests that we have moved away from
RunCLICommand will not catch a panic that we introduced here.
When using the new UsageRequirement, and an ApplicationRequirement,
or any other requirement that depends on a specific argument
id, we need to make sure to add a test that would trigger
the invalid command usage. This way the code will panic in the
test if we have not properly handled getting the arg.

[#122079997]

Signed-off-by: Jeff Pak <jeffrey.pak@emc.com>

Former-commit-id: a11fb816e27c4a754a07221ad69dd32f18d12a75 [formerly a11fb816e27c4a754a07221ad69dd32f18d12a75 [formerly 564b747c6ff8d378769c40d3803eeafaf34ce550]]
Former-commit-id: f647c3d9883a818d7ec966ce39adab8d149fcf1e
Former-commit-id: 02305d0efab98dce4af8bc057b8be1bbd257af65