properly handle '.*' in .cfignore
We were skipping the entire app directory because '.*' matches the
unexpanded '.' (this directory). When walking the directory we now check
the full path against the app directory path first and continue if it
matches, instead of skipping the entire directory. The test just makes
sure that the app directory is not skipped (it still gets other files to
upload), but dotfiles are ignored when '.*' is present in the cfignore.

(github issue #870) [fix #122234577]
Signed-off-by: Nick Wei <nwei@pivotal.io>

Former-commit-id: 487736ecb744f7d760964911e9aaef2061d61174 [formerly 487736ecb744f7d760964911e9aaef2061d61174 [formerly 27152dbbd8655efaefba12acb477c83ad286d389]]
Former-commit-id: 267192f99f1d144ffc6adc99304ab1fa9beddc6f
Former-commit-id: 34fe00af48967a15ab7de016e83f415d2f046fdd