trace: Redact all passwords and tokens in JSON
Extend the regex that redacts the values of sensitive keys in JSON responses
to include case-insensitive and substring matches of the key names
"password" and "token".

It's quite common to provide credentials to 12factor[0] applications using
environment variables. These can be unexpectedly leaked in the CLI output if
you run a command such as `CF_TRACE=true cf apps`, because the API endpoints
for space and application summaries always include `environment_json`.

This problem is mitigated by user-provided services[1], but the docs don't
state that it's a security feature and it would be good to make the default
use case safer.

Some notes about the specific implementation, I have:

- added `(?i)` to the regex to make it a case-insensitive match
- added `[^"]*` around the property name to match substrings
- used the capture group `$1` to preserve the entire matched property name
  in the replacement
- removed the separate function calls for `access_token`, `refresh_token`
  and `oldPassword` because these are now covered by the existing `token`
  and `password` calls with the new regex
- adjusted one of the test assertions to account for the fact that the value
  of `token_type` is now also redacted. This is a side-effect of the
  changes, but I don't think it's a bad thing

[0]: http://12factor.net/config
[1]: https://docs.cloudfoundry.org/devguide/services/user-provided.html


Former-commit-id: 86273f7ebf2ea597483ce4bd20e6af126678c212 [formerly 86273f7ebf2ea597483ce4bd20e6af126678c212 [formerly d61b35d9899550d52320a7bb781b55531863c042]]
Former-commit-id: a534119b9b405d5ff7c8345e683a4d32662a7bbd
Former-commit-id: 9c77cd6a5fd406d8a506bf273d77696f26d23697