Reset timeout in callback
The issue with the previous code was that data had to be received over the connection
for the reconnect to be considered valid.  The new logic makes use of the OnConnectCallback.

The test for the onConnect was explicitly asserting that OnConnectCallback was called
with the onConnect function directly, rather than asserting that calling the
OnConnectCallback's argument would call onConnect.  We updated the test to be a bit
less explicit about that, but still test the desired behavior.

[#131582719]

Signed-off-by: Sam Nelson <snelson@pivotal.io>

Former-commit-id: c07c1812bf294d4f59f5fbcd728c0adc9197fd28 [formerly c07c1812bf294d4f59f5fbcd728c0adc9197fd28 [formerly 153bb60a5dc1b79c9bd54ed92018692747f01a07]]
Former-commit-id: 4179187be2562475e3d3d8a1a1604af834389eef
Former-commit-id: c467b0b0f812f853ac2e213aa3bd6a44851ccff4