create new NewCloudControllerClient method and rename old one
Eventually we'll need to use the UAA client we create in
NewCloudControllerClient, so I renamed that method to NewClients and
added a new NewCloudControllerClient that can also be used in the API
command. This will help enforce that we do the same thing whenever we
create a Cloud Controller client.

[#127286907]


Former-commit-id: 50fc293cb2313f8aed4cf2becc8115cb0ffca6eb [formerly 50fc293cb2313f8aed4cf2becc8115cb0ffca6eb [formerly 9b6d690556741e2a88ac08c5f6b27e27e6bbceb9]]
Former-commit-id: d87407b951927921d309ab87b3cde4828ad8eee4
Former-commit-id: 1ad8ad1dc6f4b539f5337797929cc11d593e0283