create new NewCloudControllerClient method and rename old one
Eventually we'll need to use the UAA client we create in
NewCloudControllerClient, so I renamed that method to NewClients and
added a new NewCloudControllerClient that can also be used in the API
command. This will help enforce that we do the same thing whenever we
create a Cloud Controller client.

[#127286907]


Former-commit-id: 9b6d690556741e2a88ac08c5f6b27e27e6bbceb9