mention how to get invited to CF slack in CONTRIBUTING doc
The CONTRIBUTING document recommends checking in with the team before
starting work on a pull request either via a Github issue or on the #cli
channel on CF slack.

However, first-time contributors may be unaware of the process for getting an
invite to the Cloud Foundry slack team. This commit adds a breadcrumb
for people to follow.


Former-commit-id: cee75a7064652064bcdad6c928b1c8a365823ea0 [formerly cee75a7064652064bcdad6c928b1c8a365823ea0 [formerly 4c86e255ad0b764a4e78449287367a2e5a70f6f9]]
Former-commit-id: 37b47f47c5fa408eb1de1bc00ecceafb443a1526
Former-commit-id: cc101a0e4aeb5b37599b59e1e2dfe87554fa5e0c