copy cats into gopath instead of using path: in task yaml
Using path: makes the cf-acceptance-tests in the gopath a symlink on
Concourse 3.0.1. When running ginkgo with multiple nodes on Windows
some of the nodes will randomly fail with:

chdir .: The process cannot access the file because it is being
used by another process.

Running the same binary on multiple ginkgo nodes does not seem to work
if the test cmd.Dir is a symlink.


Former-commit-id: b21ca1c61457643059d666cfaed8672157058f94 [formerly b21ca1c61457643059d666cfaed8672157058f94 [formerly b7a2ef3ca68d90038d89d59cfeaefd8e2ed9b4f3]]
Former-commit-id: 959933754321e8a22a59cda34c2dc702e846a122
Former-commit-id: ae8998c2147c9a95034678e556cc74a0c807b44d