combined the GetService and GetServices tests in integration/plugin
- pushing the service broker on windows sometimes causes an IO error
(The process cannot access the file because it is being used by another
process.). We combined these two tests because they push the same
resource, and we suspect that the IO error occurs when they both try to
read the resource at the same time. Now it only needs to do it once.

[finishes #147858787]

Signed-off-by: Alex Melnik <aleksandr.melnik@suse.com>

Former-commit-id: dbc785ae19ed1cffe1a543d899e0f70aa72cb48a