combined the GetService and GetServices tests in integration/plugin
- pushing the service broker on windows sometimes causes an IO error
(The process cannot access the file because it is being used by another
process.). We combined these two tests because they push the same
resource, and we suspect that the IO error occurs when they both try to
read the resource at the same time. Now it only needs to do it once.

[finishes #147858787]

Signed-off-by: Alex Melnik <aleksandr.melnik@suse.com>

Former-commit-id: b0b20a538c9bf8860193ca6de10732bb235e57d9 [formerly b0b20a538c9bf8860193ca6de10732bb235e57d9 [formerly dbc785ae19ed1cffe1a543d899e0f70aa72cb48a]]
Former-commit-id: 86b28a875511902f86d41fd86331f389088401cd
Former-commit-id: c8f820603464a70bee8a52b842d2c4ba64b464c8