handle v2/v3 switch between app and scale commands
- The v3/shared#NewClients method adds an additional argument that can
switch the API not available error to a v2/v3 switch error.
- The v2/v3 switch error flows to the top of main#Main that will switch
between the master command list and the (limited) v2 command list when
the error is read. (Causing a full reparse of the CLI args)
- Assuming that v2's args/flags are a subset of v3's args/flags - the
addition of any new v3 specific args/flags will need to additionally be
added to the v2 version of the command.
- During the execute of the v2 command, additional validations need to
be added for v3 specific args/flags; when they are provided, the v2 command
should error out with a MinimumAPIVersionNotMetError.

Signed-off-by: Nick Wei <nwei@pivotal.io>

Former-commit-id: 1e37a8c87b69998fbb17e4fb4c11e8f21379ee59